pass eager initialized modules directly to bridge (#36916)
Summary:
Pull Request resolved: https://github.com/facebook/react-native/pull/36916

Changelog: [iOS][Removed]

the bridge only behavior of eager initialization of native modules was coupled with the turbomodule infra, even though it didn't need to be. this eager initialized modules is a static list, and the bridge owner which is usually app scoped, can pass this list directly to the bridge instead.

Reviewed By: sammy-SC

Differential Revision: D45021784

fbshipit-source-id: f033661e0722f65ae971bd7ce27f8dc7de1173ad