Use stable sort for performance entries (#36998)
Summary:
Pull Request resolved: https://github.com/facebook/react-native/pull/36998

For performance entries that have the same start/end time it makes more sense to report them back to `PerformanceObserver` in the same order they were logged.

This kind of determinism is arguably better both in terms of API, and from the point of view of testing.

Changelog: [Internal]

Reviewed By: javache

Differential Revision: D45142500

fbshipit-source-id: 77ff0093bead45dc2f15efc4b903dc181927565a